move trait defs out of mod.rs
